is_prime.cpp

+++ templates/number_theory/is_prime_int.cpp
@@ -6,9 +6,9 @@
 using namespace std;
 
 
-bool isPrime (unsigned long long n){
+bool isPrime (int n){
     if (n == 0 || n == 1) return false;
-    for (unsigned long long i=2; i <= sqrtl(n); ++i){
+    for (int i=2; i <= sqrtl(n); ++i){
         if (n % i == 0){
             return false;
         }
@@ -19,7 +19,7 @@ bool isPrime (unsigned long long n){
 
 int main(){
 
-    unsigned long long n = 9;
+    int n = 9;
 
     cout << isPrime(n) << endl;
